Moved D/I sinternals to new core package
