+++ core/src/main/scala/chisel3/experimental/hierarchy/package.scala
@@ -45,4 +45,13 @@ package object hierarchy {
     * }}}
     */
   class public extends chisel3.internal.public
+
+  type Instance[P] = core.Instance[P]
+  val Instance = core.Instance
+  type Definition[P] = core.Definition[P]
+  val Definition = core.Definition
+  type Hierarchy[P] = core.Hierarchy[P]
+  val Hierarchy = core.Hierarchy
+  type IsInstantiable = core.IsInstantiable
+  type IsLookupable = core.IsLookupable
 }
